refactor signup endpoints into separate file
